dtoweapondata

+++ UnityDataImporter/Models/WeaponData.cs
@@ -5,7 +5,7 @@ public class WeaponData
     public long Id { get; set; }
     public string ItemId { get; set; } = string.Empty;
     public long? Damage { get; set; }
-    public int? Heaviness { get; set; }
+    public float? Heaviness { get; set; }
     public string? Ammo { get; set; }
     public float? Cooldown { get; set; }
